+++ src/style/design-tokens/carbon-global-tokens-provider/index.js
@@ -1,0 +1,1 @@
+export { default } from "./carbon-global-tokens-provider.component";